refacto aria
